first modification of dependencies for melodic

+++ installation/installation_dep_script.sh
@@ -41,8 +41,8 @@ then
 	cd /tmp/driver_common/src
 	git clone https://github.com/ros-drivers/driver_common
 	cd ..
-	catkin_make install -DCMAKE_INSTALL_PREFIX=/tmp/ros/kinetic
-	sudo cp -R /tmp/ros/kinetic /opt/ros/
+	catkin_make install -DCMAKE_INSTALL_PREFIX=/tmp/ros/$ROS_DISTRO
+	sudo cp -R /tmp/ros/$ROS_DISTRO /opt/ros/
 	rm -rf /tmp/ros
 	rm -rf /tmp/driver_common
 
@@ -53,8 +53,35 @@ then
 	cd /tmp/keyboard/src
 	git clone https://github.com/lrse/ros-keyboard.git
 	cd ..
-	catkin_make install -DCMAKE_INSTALL_PREFIX=/tmp/ros/kinetic
-	sudo cp -R /tmp/ros/kinetic /opt/ros/
+	catkin_make install -DCMAKE_INSTALL_PREFIX=/tmp/ros/$ROS_DISTRO
+	sudo cp -R /tmp/ros/$ROS_DISTRO /opt/ros/
+	rm -rf /tmp/ros
+	rm -rf /tmp/keyboard
+fi
+
+if [ "$ROS_DISTRO" = "melodic" ]
+then
+	echo "----------------------------"
+	echo "Install driver common required in melodic"
+	echo "----------------------------"
+	mkdir -p /tmp/driver_common/src
+	cd /tmp/driver_common/src
+	git clone https://github.com/ros-drivers/driver_common
+	cd ..
+	catkin_make install -DCMAKE_INSTALL_PREFIX=/tmp/ros/$ROS_DISTRO
+	sudo cp -R /tmp/ros/$ROS_DISTRO /opt/ros/
+	rm -rf /tmp/ros
+	rm -rf /tmp/driver_common
+
+	echo "----------------------------"
+	echo "Install keyboard required in melodic"
+	echo "----------------------------"
+	mkdir -p /tmp/keyboard/src
+	cd /tmp/keyboard/src
+	git clone https://github.com/lrse/ros-keyboard.git
+	cd ..
+	catkin_make install -DCMAKE_INSTALL_PREFIX=/tmp/ros/$ROS_DISTRO
+	sudo cp -R /tmp/ros/$ROS_DISTRO /opt/ros/
 	rm -rf /tmp/ros
 	rm -rf /tmp/keyboard
 fi
